--- a/src/lib/CMakeLists.txt
+++ b/src/lib/CMakeLists.txt
@@ -491,7 +491,6 @@ set(
     utils/assert.hpp
     utils/boost_default_memory_resource.cpp
     utils/copyable_atomic.hpp
-    utils/cuckoo_hashtable.hpp
     utils/enum_constant.hpp
     utils/filesystem.hpp
     utils/format_bytes.cpp