--- a/src/test/CMakeLists.txt
+++ b/src/test/CMakeLists.txt
@@ -149,7 +149,6 @@ set(
     tasks/operator_task_test.cpp
     testing_assert.cpp
     testing_assert.hpp
-    utils/cuckoo_hashtable_test.cpp
     utils/format_bytes_test.cpp
     utils/format_duration_test.cpp
     utils/numa_memory_resource_test.cpp