This PR changes the hash table used in the hash join to a simple `std::unordered_map`.

The main reason was that the concurrent WIP PR #1049 triggered the sometimes occurring problems with our cuckoo table much more frequently.

On top of that and for me also surprisingly, the `std::unordered_map` is even faster:
```
Master:
./cmake_release/hyriseMicroBenchmarks --benchmark_filter=JoinHash\* --benchmark_repetitions=10
-----------------------------------------------------------------------------
Benchmark                                Time                 CPU Iterations
-----------------------------------------------------------------------------
BM_Join_Small<JoinHash>_mean       1,697,778 ns    1,689,193 ns        387
BM_Join_Skewed<JoinHash>_mean    773,188,627 ns  768,600,600 ns          1
BM_Join_Big<JoinHash>_mean        93,289,491 ns   92,909,533 ns          6

This PR:
-----------------------------------------------------------------------------
Benchmark                                Time                 CPU Iterations
-----------------------------------------------------------------------------
BM_Join_Small<JoinHash>_mean       1,494,822 ns    1,480,951 ns        464
BM_Join_Skewed<JoinHash>_mean    613,261,750 ns  609,473,200 ns          1
BM_Join_Big<JoinHash>_mean        85,425,641 ns   85,003,443 ns          7
```

This PR removes a hand-written hash table with a std-based hash table, which allows us to test other hash map implementations (sharing the same interface) that are (i) faster than std::unordered_map and (ii) more space-efficient.