@mrks 2018-08-13
Could you please also provide numbers for TPC-H?