@Bouncner 2018-08-13
And results for `scale=0.1`:
```
+-----------+---------------------+------+--------------------+------+--------+-----------------------------------------------+
| Benchmark | prev. iter/s        | runs | new iter/s         | runs | change |               p-value (significant if <0.001) |
+-----------+---------------------+------+--------------------+------+--------+-----------------------------------------------+
| TPC-H 1   | 2.4382209777832     | 13   | 2.46756267547607   | 13   | +1%    |                   (run time too short) 0.6414 |
| TPC-H 2   | 13.8464708328247    | 70   | 13.652416229248    | 69   | -1%    |                   (run time too short) 0.3529 |
| TPC-H 3   | 13.7707166671753    | 69   | 13.8171052932739   | 70   | +0%    |                   (run time too short) 0.8336 |
| TPC-H 4   | 0.327766120433807   | 2    | 0.412352234125137  | 3    | +26%   | (run time too short) (not enough runs) 0.0034 |
| TPC-H 5   | 9.81660842895508    | 50   | 9.79790782928467   | 49   | -0%    |                   (run time too short) 0.9079 |
| TPC-H 6   | 40.3500022888184    | 202  | 39.9615631103516   | 200  | -1%    |                   (run time too short) 0.2987 |
| TPC-H 7   | 2.90275120735168    | 15   | 2.97938919067383   | 15   | +3%    |                   (run time too short) 0.1686 |
| TPC-H 8   | 2.4810950756073     | 13   | 2.31846404075623   | 12   | -7%    |                   (run time too short) 0.0036 |
| TPC-H 9   | 1.43333959579468    | 8    | 1.34580373764038   | 7    | -6%    | (run time too short) (not enough runs) 0.0122 |
| TPC-H 10  | 9.56744575500488    | 48   | 9.83222484588623   | 50   | +3%    |                   (run time too short) 0.0928 |
| TPC-H 11  | 29.9462070465088    | 150  | 28.6363754272461   | 144  | -4%    |                   (run time too short) 0.0000 |
| TPC-H 12  | 13.2450475692749    | 67   | 13.6787977218628   | 69   | +3%    |                   (run time too short) 0.0066 |
| TPC-H 13  | 12.8717784881592    | 65   | 11.2632703781128   | 57   | -12%   |                   (run time too short) 0.0000 |
| TPC-H 14  | 28.742431640625     | 144  | 29.2504615783691   | 147  | +2%    |                   (run time too short) 0.0556 |
| TPC-H 15  | 0.215460941195488   | 2    | 0.223195970058441  | 2    | +4%    | (run time too short) (not enough runs) 0.3505 |
| TPC-H 16  | 6.56419467926025    | 33   | 6.17109727859497   | 31   | -6%    |                   (run time too short) 0.0100 |
| TPC-H 17  | 1.32866430282593    | 7    | 1.30945432186127   | 7    | -1%    | (run time too short) (not enough runs) 0.2780 |
| TPC-H 18  | 0.0222702454775572  | 1    | 0.023688118904829  | 1    | +6%    |    (run time too short) (not enough runs) nan |
| TPC-H 19  | 1.33674013614655    | 7    | 1.2693567276001    | 7    | -5%    | (run time too short) (not enough runs) 0.3823 |
| TPC-H 20  | 0.00526655884459615 | 1    | 0.0053626848384738 | 1    | +2%    |                         (not enough runs) nan |
| TPC-H 21  | 0.230141177773476   | 2    | 0.234685733914375  | 2    | +2%    | (run time too short) (not enough runs) 0.6825 |
| TPC-H 22  | 0.290742963552475   | 2    | 0.295315533876419  | 2    | +2%    | (run time too short) (not enough runs) 0.4553 |
| average   |                     |      |                    |      | +0%    |                                               |
+-----------+---------------------+------+--------------------+------+--------+-----------------------------------------------+
```