@mrks 2018-08-14
Other than the missing `using`, I don't have anything to comment on the code side. As the TPC-H benchmarks are way less convincing than the previous benchmarks, I wonder if we should discuss this tomorrow.