@Bouncner 2018-08-14
I will run some longer running TPC-H benchmarks to get a more accurate picture. Is nemea the best place for that?