@Bouncner 2018-08-15
Here a longer running TPC-H benchmark. Scale factor 0.1.

```
+-----------+---------------------+------+---------------------+------+--------+---------------------------------+
| Benchmark | prev. iter/s        | runs | new iter/s          | runs | change | p-value (significant if <0.001) |
+-----------+---------------------+------+---------------------+------+--------+---------------------------------+
| TPC-H 1   | 2.73134922981262    | 1000 | 2.70583200454712    | 1000 | -1%    |                          0.0000 |
| TPC-H 2   | 20.9591464996338    | 1000 | 22.0250205993652    | 1000 | +5%    |     (run time too short) 0.0000 |
| TPC-H 3   | 16.2167778015137    | 1000 | 22.2300300598145    | 1000 | +37%   |     (run time too short) 0.0000 |
| TPC-H 4   | 0.31084680557251    | 224  | 0.310456722974777   | 224  | -0%    |                          0.5524 |
| TPC-H 5   | 12.1663751602173    | 1000 | 15.6403093338013    | 1000 | +29%   |                          0.0000 |
| TPC-H 6   | 49.6075668334961    | 1000 | 49.271183013916     | 1000 | -1%    |     (run time too short) 0.0074 |
| TPC-H 7   | 4.29073572158813    | 1000 | 4.75418281555176    | 1000 | +11%   |                          0.0000 |
| TPC-H 8   | 3.10401010513306    | 1000 | 3.23372077941895    | 1000 | +4%    |                          0.0000 |
| TPC-H 9   | 1.78640758991241    | 1000 | 1.92059361934662    | 1000 | +8%    |                          0.0000 |
| TPC-H 10  | 14.4318647384644    | 1000 | 15.4717359542847    | 1000 | +7%    |                          0.0000 |
| TPC-H 11  | 40.8472366333008    | 1000 | 44.0052032470703    | 1000 | +8%    |     (run time too short) 0.0000 |
| TPC-H 12  | 17.6362342834473    | 1000 | 19.3214855194092    | 1000 | +10%   |     (run time too short) 0.0000 |
| TPC-H 13  | 13.5514631271362    | 1000 | 17.6978759765625    | 1000 | +31%   |     (run time too short) 0.0000 |
| TPC-H 14  | 39.5775451660156    | 1000 | 40.5096244812012    | 1000 | +2%    |     (run time too short) 0.0000 |
| TPC-H 15  | 0.268731832504272   | 194  | 0.27754470705986    | 200  | +3%    |                          0.0000 |
| TPC-H 16  | 6.81569290161133    | 1000 | 6.89850330352783    | 1000 | +1%    |                          0.0000 |
| TPC-H 17  | 1.3637033700943     | 982  | 1.36338913440704    | 982  | -0%    |                          0.9027 |
| TPC-H 18  | 0.0334157757461071  | 25   | 0.0330488756299019  | 24   | -1%    |                          0.0000 |
| TPC-H 19  | 1.73480129241943    | 1000 | 1.76322031021118    | 1000 | +2%    |                          0.0000 |
| TPC-H 20  | 0.00561590725556016 | 5    | 0.00562030030414462 | 5    | +0%    |        (not enough runs) 0.6793 |
| TPC-H 21  | 0.184361517429352   | 133  | 0.185666546225548   | 134  | +1%    |                          0.0002 |
| TPC-H 22  | 0.306500941514969   | 221  | 0.306894689798355   | 221  | +0%    |                          0.5168 |
| average   |                     |      |                     |      | +7%    |                                 |
+-----------+---------------------+------+---------------------+------+--------+---------------------------------+
```

Max loses of -1% (pretty sure it's not that PR) and improvements of up to 37%.