@mrks 2018-08-15
What's up with the `murmur_hash.*` files? Do we still need those?