@Bouncner 2018-08-15
We use murmur hashing for the radix cluster calculation (does not really hurt to use std::hash here for our simple values). So it's used but could be removed.

Only thing that might use murmur in the future are the filters from @ArneMayer . Can you tell us something about that, @ArneMayer ?