@Bouncner 2018-08-15
Whooopsi ... it's new branch.
I just pushed the changes with `using ...`. Shall I open a new PR for that? It's just 4 lines ...