@ArneMayer 2018-08-17
@Bouncner The filters bring their own hashing function.